pyproject.toml
[project]
name = "lifesync-api"
version = "0.0.1.dev1"
description = ""
authors = [
    {name = "Henrique Sebastião",email = "contato@henriquesebastiao.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.7"
taskipy = "^1.14.1"
radon = "^6.0.1"
bandit = "^1.8.6"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[tool.ruff]
line-length = 99
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff format .; ruff check . --fix'
run = 'poetry run fastapi dev lifesync/main.py'
radon = 'radon cc ./lifesync -a -na'
bandit = 'bandit -r ./lifesync'
export = './scripts/rm-requirements.sh && poetry export -f requirements.txt --output requirements.txt --without-hashes --without dev'